1.0.5 automatically generated from 053b420

+++ lib/swagger_client/models/audit_log_entry.rb
@@ -57,7 +57,7 @@ module SwaggerClient
       {
         :'_links' => :'Links',
         :'_id' => :'String',
-        :'date' => :'Float',
+        :'date' => :'Integer',
         :'kind' => :'String',
         :'name' => :'String',
         :'description' => :'String',
